# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export type OrderInputKpData = { [key: string]: unknown };

export interface OrderInput {
  /** @minLength 1 */
  prefix: string;
  /** @minLength 1 */
  zoneLabel: string;
  kpData: OrderInputKpData;
}

export interface OrderCreated {
  id: number;
  orderNumber: string;
  pdfToken: string;
  [key: string]: unknown;
 }

export interface OrderPdfPermission {
  /** @maxLength 2048 */
  pdfToken: string;
}

export interface OrderPdfUpload {
  uploadURL: string;
  objectPath: string;
  uploadToken: string;
}

export interface OrderPdfAttachment {
  /** @pattern ^/objects/uploads/[a-f0-9-]{36}$ */
  objectPath: string;
  /** @maxLength 2048 */
  uploadToken: string;
}

export interface OrderPdfSaved {
  ok: boolean;
}

export interface HealthStatus {
  status: string;
}