# lib/api-zod/src/generated/types/orderCreated.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderCreated {
  id: number;
  orderNumber: string;
  pdfToken: string;
  [key: string]: unknown;
 }