# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputKpData } from './orderInputKpData';

export interface OrderInput {
  /** @minLength 1 */
  prefix: string;
  /** @minLength 1 */
  zoneLabel: string;
  kpData: OrderInputKpData;
}