# lib/api-zod/src/generated/types/orderPdfAttachment.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderPdfAttachment {
  /** @pattern ^/objects/uploads/[a-f0-9-]{36}$ */
  objectPath: string;
  /** @maxLength 2048 */
  uploadToken: string;
}