# lib/api-zod/src/generated/types/orderPdfPermission.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderPdfPermission {
  /** @maxLength 2048 */
  pdfToken: string;
}