# lib/api-zod/src/generated/types/orderPdfUpload.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderPdfUpload {
  uploadURL: string;
  objectPath: string;
  uploadToken: string;
}